backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
selenium==4.25.0
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
gspread==6.1.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
thefuzz==0.22.1
python-Levenshtein==0.25.1
pydantic==2.9.2
python-dotenv==1.0.1
